changed BigDecimal to Double since GAE apparently doesn't support BD ugh

+++ SDCrimeZone-AppEngine/src/com/crimezone/sd/server/domain/Incident.java
@@ -1,6 +1,5 @@
 package com.crimezone.sd.server.domain;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Entity;
@@ -17,8 +16,8 @@ public class Incident {
   private Date date;
   private String bccCode;
   private String address;
-  private BigDecimal latitude;
-  private BigDecimal longitude;
+  private Double latitude;
+  private Double longitude;
   private Integer year; // used for queries
 
   public Incident() {
@@ -56,19 +55,19 @@ public class Incident {
     this.address = address;
   }
 
-  public BigDecimal getLatitude() {
+  public Double getLatitude() {
     return latitude;
   }
 
-  public void setLatitude(BigDecimal latitude) {
+  public void setLatitude(Double latitude) {
     this.latitude = latitude;
   }
 
-  public BigDecimal getLongitude() {
+  public Double getLongitude() {
     return longitude;
   }
 
-  public void setLongitude(BigDecimal longitude) {
+  public void setLongitude(Double longitude) {
     this.longitude = longitude;
   }
